Clean up common mock class and how it's inherited
